% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataextr.number.R
\name{dataextr.number}
\alias{dataextr.number}
\title{Number of records remaining after full data extraction}
\usage{
dataextr.number(ca.number = 76.12213, dextraction.retained = 100)
}
\description{
This function calculates the number of records remaining after full data
extraction (i.e. study qualitative or quanitative fundings) in a systematic
review, based on the inputs of the number of records remaining after critical
appraisal ('ca.number', see 'ca.number' function) and the percentage of records
retained through this full data extraction stage ('dextraction.retained'). Default
values are provided based on the empirical study of environmental systematic
reviews by Haddaway and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
