% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataextr.time.R
\name{dataextr.time}
\alias{dataextr.time}
\title{Time needed to extract full data from all included records}
\usage{
dataextr.time(ca.number = 76.12213, dextraction.day = 6.9,
  dextraction.checked = 0)
}
\description{
This function calculates the time needed to extract full data
(i.e. study qualitative or quantitative findings) from all
relevant records in a systematic review, based on the inputs of
the number of records passing the critical appraisal stage
('ca.number', see 'ca.number' function), the
number of studies from which full data can be extracted per day
('dextraction.day'), and the percentage of all studies that are
double checked for consistency of full data extraction
('dextraction.checked'). Where full dual screening of all
records is used, this will equal a percentage of 100 abstracts
being checked. Default values are provided based on the
empirical study of environmental systematic reviews by
Haddaway and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
