% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.time.R
\name{dataprep.time}
\alias{dataprep.time}
\title{Time needed to prepare extracted data from all included records}
\usage{
dataprep.time(dataextr.number = 76.12213, dprep.day = 24,
  dprep.checked = 0)
}
\description{
This function calculates the time needed to prepare extracted data
for synthesis (i.e. calculating effect sizes and variability estimates,
like standard error) from all relevant records in a systematic review,
based on the inputs of the number of records passing the full data
extraction stage ('dataextr.number', see 'dataextr.number' function), the
number of studies whose data can be prepared per day ('dprep.day'),
and the percentage of all studies that are double checked for consistency
of data preparation ('dprep.checked'). Where full dual screening of all
records is used, this will equal a percentage of 100 abstracts
being checked. Default values are provided based on the
empirical study of environmental systematic reviews by
Haddaway and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
