% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclabstracts.number.R
\name{inclabstracts.number}
\alias{inclabstracts.number}
\title{Number of abstracts included after abstract screening}
\usage{
inclabstracts.number(incltitles.number = 1240.665,
  abstracts.retained = 25)
}
\description{
This function calculates the number of abstracts remaining after abstract
screening in a systematic review, based on the inputs of the number of
records remaining after title screening ('incltitles.number', see
'incltitles.number' function) and the percentage of abstracts retained
through this abstract screening stage ('abstracts.retained'). Default
values are provided based on the empirical study of environmental
systematic reviews by Haddaway and Westgate (2018)
https://doi.org/10.1111/cobi.13231.
}
