% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclft.number.R
\name{inclft.number}
\alias{inclft.number}
\title{Number of full texts included after full text screening}
\usage{
inclft.number(retrieval.number = 468.0408, ft.retained = 21.4)
}
\description{
This function calculates the number of full texts remaining after full text
screening in a systematic review, based on the inputs of the number of
records remaining after retrieval ('retrieval.number', see
'retrieval.number' function) and the percentage of full texts retained
through this full text screening stage ('ft.retained'). Default values are
provided based on the empirical study of environmental systematic reviews
by Haddaway and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
