% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incltitles.number.R
\name{incltitles.number}
\alias{incltitles.number}
\title{Number of titles included after title screening}
\usage{
incltitles.number(uniqart.number = 8497.706, titles.retained = 14.6)
}
\description{
This function calculates the number of titles remaining after title
screening in a systematic review, based on the inputs of the number of
unique articles after duplicate removal ('uniqart.number', see
'uniqart.number' function) and the percentage of titles retained
through this title screening stage ('titles.retained'). Default values
are provided based on the empirical study of environmental systematic
reviews by Haddaway and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
