% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdataextr.number.R
\name{mdataextr.number}
\alias{mdataextr.number}
\title{Number of records remaining after meta-data extraction}
\usage{
mdataextr.number(inclft.number = 100.1607, mdextraction.retained = 100)
}
\description{
This function calculates the number of records remaining after meta-data
extraction in a systematic review, based on the inputs of the number of
records remaining after full text screening ('inclft.number', see
'inclft.number' function) and the percentage of records retained
through this meta-data extraction stage ('mdextraction.retained'). Default
values are provided based on the empirical study of environmental systematic
reviews by Haddaway and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
