% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Systematic review sample data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://doi.org/10.1111/cobi.13231}
}
\usage{
data(sampledata)
}
\description{
Dataset comes from an analysis of systematic reviews (SRs) and systematic maps (SM) published by the
Collaboration for Environmental Evidence (CEE, www.environmentalevidence.org) between May 2012 and
March 2017. Data are based on the empirical study of environmental systematic reviews by Haddaway
and Westgate (2018) https://doi.org/10.1111/cobi.13231.
}
\examples{
data(sampledata)
str(sampledata)
}
\references{
Haddaway and Westgate (2018) Conservation Biology, 33(2):434-443
(\href{https://doi.org/10.1111/cobi.13231})
}
\keyword{datasets}
