% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecalc.R
\name{timecalc}
\alias{timecalc}
\title{Calculate time needed for a systematic review}
\usage{
timecalc(abstracts.checked, abstracts.day, abstracts.retained, add.search,
  administration, ca.checked, ca.day, ca.retained, comms, databasesdb.day,
  deduping, dextraction.checked, dextraction.day, dextraction.retained,
  dprep.checked, dprep.day, ft.checked, ft.day, ft.retained, fte,
  greylit.searching, greylitgreylit.day, mdextraction.checked,
  mdextraction.day, mdextraction.retained, meetings, orig.articles,
  percent.unique, planning, protocol, reflists, reportwriting,
  retrieval.checked, retrieval.day, retrieval.retained, synthesis,
  titles.checked, titles.day, titles.retained)
}
\description{
This function calculates the time needed to conduct a systematic review, based on
a range of input data (abstracts.checked,abstracts.day,abstracts.retained,add.search,
administration,ca.checked,ca.day,ca.retained,comms,databasesdb.day,deduping,
dextraction.checked,dextraction.day,dextraction.retained,dprep.checked,dprep.day,
ft.checked,ft.day,ft.retained,fte,greylit.searching,greylitgreylit.day,
mdextraction.checked,mdextraction.day,mdextraction.retained,meetings,orig.articles,
percent.unique,planning,protocol,reflists,reportwriting,retrieval.checked,
retrieval.day,retrieval.retained,synthesis,titles.checked,titles.day,titles.retained).
}
