% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usedefaults.R
\name{usedefaults}
\alias{usedefaults}
\title{Impute missing values in input PredicTER dataset using defaults}
\usage{
usedefaults(planning, protocol, administration, fte, databases, db.day,
  greylit, greylit.day, reflists, add.search, orig.articles,
  percent.unique, deduping, titles.day, titles.retained, titles.checked,
  abstracts.day, abstracts.retained, abstracts.checked, ta.day,
  ta.retained, ta.checked, retrieval.day, retrieval.retained,
  retrieval.checked, ft.day, ft.retained, ft.checked, mdextraction.day,
  mdextraction.retained, mdextraction.checked, ca.day, ca.retained,
  ca.checked, dextraction.day, dextraction.retained, dextraction.checked,
  dprep.day, dprep.retained, dprep.checked, synthesis, reportwriting,
  comms, meetings)
}
\description{
This function takes an input dataset and imputes missing values for specific columns using
empirically-derived default values. Default values are provided based on the empirical
study of environmental systematic reviews by Haddaway and Westgate (2018)
https://doi.org/10.1111/cobi.13231.
}
